fix image deletion

+++ grails-app/jobs/ims/DeleteImageFileJob.groovy
@@ -5,6 +5,10 @@ import be.cytomine.client.collections.DeleteCommandCollection
 import be.cytomine.client.models.DeleteCommand
 import grails.converters.JSON
 import org.codehaus.groovy.grails.web.json.JSONElement
+import be.cytomine.formats.FormatIdentifier
+import be.cytomine.formats.supported.digitalpathology.OpenSlideMultipleFileFormat
+import be.cytomine.formats.heavyconvertable.CellSensVSIFormat
+import be.cytomine.exception.FormatException
 
 
 class DeleteImageFileJob {
@@ -36,9 +40,41 @@ class DeleteImageFileJob {
 
             File fileToDelete = new File(j.path+j.filename)
 
-            if(fileToDelete.exists()) {
-                log.info "DELETE file "+fileToDelete.absolutePath
-                fileToDelete.delete()
+            def format
+            try{
+                format = FormatIdentifier.getImageFormat(fileToDelete.absolutePath)
+            } catch(FormatException e) {
+                if(fileToDelete.isFile()) log.error "Unkown format for "+fileToDelete.absolutePath
+                else log.info "Unkown format for "+fileToDelete.absolutePath
+            }
+
+            if(fileToDelete.exists() && format) {
+                if(!(format instanceof OpenSlideMultipleFileFormat) && !(format instanceof CellSensVSIFormat)) {
+                    log.info "DELETE file "+fileToDelete.absolutePath
+                    fileToDelete.delete()
+                } else if(format instanceof CellSensVSIFormat) {
+                    if(fileToDelete.isFile() && fileToDelete.absolutePath.endsWith(".vsi")) fileToDelete = fileToDelete.parentFile
+
+                    File vsiFile = fileToDelete.listFiles().find {it.isFile() && it.absolutePath.endsWith(".vsi")}
+                    File vsiFolder =  fileToDelete.listFiles().find {it.isDirectory() && it.absolutePath.contains(vsiFile.name.replace(".vsi",""))}
+                    log.info "DELETE file "+vsiFile.absolutePath
+                    vsiFile.delete();
+                    log.info "DELETE folder "+vsiFolder.absolutePath
+                    vsiFolder.deleteDir()
+
+                    if (fileToDelete.listFiles().size() == 0){
+                        log.info "DELETE folder "+fileToDelete.absolutePath
+                        fileToDelete.delete()
+                    }
+                } else if(format instanceof OpenSlideMultipleFileFormat) {
+                    if(fileToDelete.isFile()) fileToDelete = fileToDelete.parentFile
+                    log.info "DELETE folder "+fileToDelete.absolutePath
+                    fileToDelete.deleteDir()
+                }
+                if (fileToDelete.parentFile.listFiles().size() == 0){
+                    log.info "DELETE folder "+fileToDelete.parentFile.absolutePath
+                    fileToDelete.parentFile.delete()
+                }
             }
 
         }
